Fix key access to possibly non-existent location
The location may be non-existent in the returned data. This is correctly
handled in the main code, but not in the logging entries, possibly
breaking the data update of the integration.

This change also simplifies the handling of a missing location in the
main code flow to have it consistent between the locations.